Breadcrumb spike, no arrows yet.
It is 'just' a nav with links, so could be styled more simply if desired.

The black bar on top is getting a bit big, so black on white might be better
than white on dark grey.

More grid. it is very clear where you are - it seems. The bold in the menu for
'current' can probably go.