Don't use asset pipeline just yet.
Not supported on github. Left the gemfiles and the logos in the asset
pipeline. Setting up a build would be better with a pair.

I did learn that apparently the DSL for 'travis' builds is open source,
and supported by 20+ platforms (including Hackage from the list of
contributors).

Version 2 is out, but couldn't find a copy-pastable instruction. So I
expect some yak shaving, might be better to try this out with a clean
site first.